# ApnaStayBackEnd/docs/swagger.json
  {
    "openapi": "3.0.3",
    "info": {
      "title": "HomeFinity Backend API",
      "description": "API for HomeFinity frontend.\n\n**Required headers for protected endpoints:**\n- **Authorization** (mandatory): `Authorization: Bearer <JWT>` — JWT from POST /api/auth/public/signin or POST /api/auth/public/phone/verify-and-login.\n- **X-XSRF-TOKEN** (mandatory): CSRF token from GET /api/csrf-token.\n\n**Required header for requests with body (POST/PUT):**\n- **Content-Type** (mandatory): `Content-Type: application/json`.\n\nPublic endpoints (signin, signup, phone send-code, phone verify-and-login, forgot-password, reset-password, verify-2fa-login, csrf-token, health-check) do not require Authorization or X-XSRF-TOKEN.",
      "version": "0.2.0"
    },
    "servers": [
      {
        "url": "/",
        "description": "Relative to deployed base URL"
      }
    ],
...
    "security": [
      {
        "bearerAuth": [],
        "csrfHeader": []
      }
    ],
    "paths": {
...
      "/api/auth/username": {
        "get": {
          "tags": ["Auth"],
          "summary": "Get current username",
          "description": "Returns the authenticated user's username. Requires Authorization and X-XSRF-TOKEN (mandatory).",
          "operationId": "getUsername",
"parameters": [
          { "$ref": "#/components/parameters/XsrfTokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "Username or error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": { "type": "boolean" },
                      "message": { "type": "string" },
                      "timestamp": { "type": "string", "format": "date-time" },
                      "data": { "type": "string", "description": "Username when success is true" }
                    }
                  }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        }
      },
...
      "/api/admin/profiles/{role}/{id}/approve": {
        "put": {
          "tags": ["Admin"],
          "summary": "Approve profile",
          "description": "Approve a profile (admin only). Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "approveProfile",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "role", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/AppRole" } },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
            { "name": "adminNote", "in": "query", "required": false, "schema": { "type": "string" } }
          ],
          "requestBody": {
            "required": false,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApproveRejectRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Profile approved",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseVoid" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" },
            "404": { "description": "Profile not found" }
          }
        }
      },
      "/api/admin/profiles/{role}/{id}/reject": {
        "put": {
          "tags": ["Admin"],
          "summary": "Reject profile",
          "description": "Reject a profile (admin only). Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "rejectProfile",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "role", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/AppRole" } },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
            { "name": "adminNote", "in": "query", "required": false, "schema": { "type": "string" } }
          ],
          "requestBody": {
            "required": false,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApproveRejectRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Profile rejected",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseVoid" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" },
            "404": { "description": "Profile not found" }
          }
        }
      },
...
      "/api/profile/roles": {
        "get": {
          "tags": ["Profile"],
          "summary": "Get profile roles",
          "description": "Returns available profile roles (OWNER, BROKER, USER). Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "getProfileRoles",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" }
          ],
          "responses": {
            "200": {
              "description": "List of profile roles",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseAppRoleList" }
                }
              }
            },
            "401": { "description": "Unauthorized" }
          }
        }
      },
      "/api/profile/list": {
        "get": {
          "tags": ["Profile"],
          "summary": "List all profiles (admin)",
          "description": "Returns all profiles with details, optionally filtered by role. Admin only. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "listAllProfilesWithDetails",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "role", "in": "query", "required": false, "schema": { "$ref": "#/components/schemas/AppRole" } }
          ],
          "responses": {
            "200": {
              "description": "List of profiles",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseProfileListFull" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" }
          }
        }
      },
      "/api/profile": {
        "get": {
          "tags": ["Profile"],
          "summary": "Get my profile",
          "description": "Returns the current user's profile for the given role. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "getMyProfile",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "role", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/AppRole" } }
          ],
          "responses": {
            "200": {
              "description": "Profile details",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseProfile" }
                }
              }
            },
            "400": { "description": "Missing role parameter" },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" }
          }
        },
        "put": {
          "tags": ["Profile"],
          "summary": "Update profile",
          "description": "Update the current user's profile. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "updateProfile",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" }
          ],
          "requestBody": {
            "required": false,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Profile updated",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseProfile" }
                }
              }
            },
            "400": { "description": "Validation error" },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" }
          }
        }
      },
      "/api/profile/review": {
        "post": {
          "tags": ["Profile"],
          "summary": "Submit profile for review",
          "description": "Submit profile for admin approval. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "submitForReview",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileRequest" }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Profile submitted for review",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseProfile" }
                }
              }
            },
            "400": { "description": "Validation error or missing role" },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" }
          }
        }
      },
      "/api/profile/approval-status": {
        "get": {
          "tags": ["Profile"],
          "summary": "Get approval status",
          "description": "Returns the approval status of the current user's profile. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "getApprovalStatus",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "role", "in": "query", "required": true, "schema": { "$ref": "#/components/schemas/AppRole" } }
          ],
          "responses": {
            "200": {
              "description": "Approval status",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseApprovalStatus" }
                }
              }
            },
            "400": { "description": "Missing role parameter" },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" }
          }
        }
      },
      "/api/profile/{role}/{id}/approve": {
        "put": {
          "tags": ["Profile"],
          "summary": "Approve profile (admin)",
          "description": "Approve a profile by role and ID. Admin only. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "profileApprove",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "role", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/AppRole" } },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
          ],
          "requestBody": {
            "required": false,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApproveRejectRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Profile approved",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseVoid" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" },
            "404": { "description": "Profile not found" }
          }
        }
      },
      "/api/profile/{role}/{id}/reject": {
        "put": {
          "tags": ["Profile"],
          "summary": "Reject profile (admin)",
          "description": "Reject a profile by role and ID. Admin only. Requires Authorization and X-XSRF-TOKEN.",
          "operationId": "profileReject",
          "parameters": [
            { "$ref": "#/components/parameters/XsrfTokenHeader" },
            { "name": "role", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/AppRole" } },
            { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
          ],
          "requestBody": {
            "required": false,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApproveRejectRequest" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Profile rejected",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ApiResponseVoid" }
                }
              }
            },
            "401": { "description": "Unauthorized" },
            "403": { "description": "Forbidden" },
            "404": { "description": "Profile not found" }
          }
        }
      },
...
    },
    "components": {
      "parameters": {
        "XsrfTokenHeader": {
          "in": "header",
          "name": "X-XSRF-TOKEN",
          "required": true,
          "schema": {
            "type": "string",
            "description": "CSRF token from GET /api/csrf-token (use the token value from response)"
          },
          "description": "**Mandatory** for all protected endpoints. Obtain token from GET /api/csrf-token. Authorization is defined via securitySchemes (Bearer JWT). Content-Type is defined by requestBody.content for operations with a body."
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "description": "JWT token from /api/auth/public/signin. Send as: Authorization: Bearer <token>"
        },
        "csrfHeader": {
          "type": "apiKey",
          "in": "header",
          "name": "X-XSRF-TOKEN",
          "description": "CSRF token from GET /api/csrf-token. Use the header name returned by that endpoint (e.g. X-XSRF-TOKEN) and the token value. Required for state-changing requests and optionally for GETs when cookies are used."
        }
      },
      "schemas": {
...
        "ApiResponseVoid": {
          "type": "object",
          "properties": {
            "success": { "type": "boolean" },
            "message": { "type": "string" },
            "timestamp": { "type": "string", "format": "date-time" },
            "data": { "type": "object", "nullable": true }
          }
        },
...
        "AppRole": {
          "type": "string",
          "enum": ["ROLE_ADMIN", "ROLE_OWNER", "ROLE_USER", "ROLE_BROKER"]
        },
        "ProfileStatus": {
          "type": "string",
          "enum": ["PENDING", "APPROVED", "REJECTED", "IN_PROGRESS"]
        },
        "ProfileRequest": {
          "type": "object",
          "properties": {
            "role": { "$ref": "#/components/schemas/AppRole" },
            "fullName": { "type": "string", "maxLength": 100 },
            "gender": { "type": "string", "maxLength": 20 },
            "dateOfBirth": { "type": "string", "format": "date" },
            "aadharNumber": { "type": "string", "maxLength": 12 },
            "mobile": { "type": "string", "maxLength": 15 },
            "email": { "type": "string", "format": "email", "maxLength": 100 },
            "firmName": { "type": "string", "maxLength": 150 },
            "licenseNumber": { "type": "string", "maxLength": 50 },
            "idType": { "type": "string", "maxLength": 50 },
            "idNumber": { "type": "string", "maxLength": 50 },
            "address": { "type": "string", "maxLength": 255 },
            "city": { "type": "string", "maxLength": 100 },
            "state": { "type": "string", "maxLength": 100 },
            "pinCode": { "type": "string", "pattern": "^[0-9]{6}$", "maxLength": 6 }
          }
        },
        "ProfileDTO": {
          "type": "object",
          "properties": {
            "id": { "type": "integer", "format": "int64" },
            "userId": { "type": "integer", "format": "int64" },
            "userName": { "type": "string" },
            "profileRole": { "$ref": "#/components/schemas/AppRole" },
            "fullName": { "type": "string" },
            "gender": { "type": "string" },
            "dateOfBirth": { "type": "string", "format": "date" },
            "aadharNumber": { "type": "string" },
            "mobile": { "type": "string" },
            "email": { "type": "string" },
            "firmName": { "type": "string" },
            "licenseNumber": { "type": "string" },
            "idType": { "type": "string" },
            "idNumber": { "type": "string" },
            "address": { "type": "string" },
            "city": { "type": "string" },
            "state": { "type": "string" },
            "pinCode": { "type": "string" },
            "status": { "$ref": "#/components/schemas/ProfileStatus" },
            "submittedAt": { "type": "string", "format": "date-time" },
            "reviewedAt": { "type": "string", "format": "date-time" },
            "adminNote": { "type": "string" },
            "createdAt": { "type": "string", "format": "date-time" },
            "updatedAt": { "type": "string", "format": "date-time" }
          }
        },
...
        "ApprovalStatusResponse": {
          "type": "object",
          "properties": {
            "approved": { "type": "boolean" },
            "status": { "$ref": "#/components/schemas/ProfileStatus" }
          }
        },
        "ApproveRejectRequest": {
          "type": "object",
          "properties": {
            "adminNote": { "type": "string" }
          }
        },
        "ApiResponseProfile": {
          "type": "object",
          "properties": {
            "success": { "type": "boolean" },
            "message": { "type": "string" },
            "timestamp": { "type": "string", "format": "date-time" },
            "data": { "$ref": "#/components/schemas/ProfileDTO" }
          }
        },
...
        "ApiResponseProfileListFull": {
          "type": "object",
          "properties": {
            "success": { "type": "boolean" },
            "message": { "type": "string" },
            "timestamp": { "type": "string", "format": "date-time" },
            "data": { "type": "array", "items": { "$ref": "#/components/schemas/ProfileDTO" } }
          }
        },
        "ApiResponseAppRoleList": {
          "type": "object",
          "properties": {
            "success": { "type": "boolean" },
            "message": { "type": "string" },
            "timestamp": { "type": "string", "format": "date-time" },
            "data": { "type": "array", "items": { "$ref": "#/components/schemas/AppRole" } }
          }
        },
        "ApiResponseApprovalStatus": {
          "type": "object",
          "properties": {
            "success": { "type": "boolean" },
            "message": { "type": "string" },
            "timestamp": { "type": "string", "format": "date-time" },
            "data": { "$ref": "#/components/schemas/ApprovalStatusResponse" }
          }
        },
...
      }
    }
  }